+++ app/src/main/java/org/simple/clinic/bloodsugar/entry/BloodSugarEntryUi.kt
@@ -6,31 +6,8 @@ import java.time.LocalDate
 import java.util.UUID
 
 interface BloodSugarEntryUi {
-  fun setBloodSugarSavedResultAndFinish()
-  fun hideBloodSugarErrorMessage()
-  fun showBloodSugarEmptyError()
-  fun showBloodSugarHighError(
-      measurementType: BloodSugarMeasurementType,
-      unitPreference: BloodSugarUnitPreference
-  )
-
-  fun showBloodSugarLowError(
-      measurementType: BloodSugarMeasurementType,
-      unitPreference: BloodSugarUnitPreference
-  )
-
-  fun showBloodSugarEntryScreen()
-  fun showDateEntryScreen()
-  fun showInvalidDateError()
-  fun showDateIsInFutureError()
-  fun hideDateErrorMessage()
-  fun setDateOnInputFields(date: LocalDate)
-  fun showBloodSugarDate(date: LocalDate)
   fun showRemoveButton()
   fun hideRemoveButton()
-  fun setBloodSugarReading(bloodSugarReading: String)
-  fun dismiss()
-  fun showConfirmRemoveBloodSugarDialog(bloodSugarMeasurementUuid: UUID)
   fun showEntryTitle(measurementType: BloodSugarMeasurementType)
   fun showEditTitle(measurementType: BloodSugarMeasurementType)
   fun showProgress()
@@ -45,5 +22,4 @@ interface BloodSugarEntryUi {
   fun numericBloodSugarInputType()
   fun setLabelForHbA1c()
   fun setLabelForUnknown()
-  fun showBloodSugarUnitSelectionDialog(bloodSugarUnitPreference: BloodSugarUnitPreference)
 }